[MDM] Fix layering policy violations
Policy dictates that the interface between swe browser
licy dictates that the interface between swe browser
and the Engine is to be tightly controlled.  Swe browser
classes should not be importing classes from chrome or
other 'engine' classes directly -- they need to be proxied
through an interface class.

Several MDM classes were in violation of this policy, and
have beed re-worked to proxy through the new static
interface class, MdmManager.java.

Also refactored ProxyRestriction to pass data to
ProxyChangeListener via intent extras, which eliminates
the need to create proxy functions for this purpose in
MdmManager.java

Change-Id: I2dcb3c0c3fb2a28cf7339ca12c60da435bf39cde