Fix issue where bitmaps of tab are purged incorrectly
The tab index used in the key generation and deletion was based on index
of the tab in the tab manager. When a tab was closed the indices in the
tab manager is recalculated which inturn caused deletion of incorrect
bitmaps for the tab. Changed the code to use the id generated
by the tab to be used in the key generation to make it consistent.

Change-Id: I268ea2c8324f1832b0dea6d6aa63d14eca268e13