[MDM] Rework Third Party Cookie Restriction for new UI
Previous implementation of the Third Party Cookie (TPC)
Restriction was done before SWE had a UI element to
control this feature. The implementation included a
separate override for the MDM setting since it was
not clear how the TPC preference could/would be changed/set.

Note that the original implementation was functional after
the addition of the UI element because of the 'override'
design, but of course now that there is a UI element, we
need to manage it as well. Since we are now managing
the UI element, it makes sense to align the implementation
with other similar MDM implementations (e.g. Do-Not-Track)
which means we no longer need the 'override' design.

This patch removes the 'override' code, and re-implements
TPC Restrictions in the same manner as other UI bound
features.

Change-Id: I5898409d3f0aa4d5b254d00032feda1d6080bfd1