[MDM] Added restrictionEnable to AutoFill restriction
THe Chromium policy document had a single controlling key for
Autofill, and I attempted to emulate it.  Unfortunately, a side-effect
was reported in the non-mdm browser.

The Fix was to go ahead and implement a second key (as we do for other
restrictions), restrictionEnable.  As a bonus, the enable logic for the actual
value is no longer reversed, therfore les confusing.

Change-Id: Ib074bf3efa79bded3fca246bf63824d640c26954